added login and register in clean archi

+++ server-side/use-cases/questions/remove-question.js
@@ -3,7 +3,9 @@ export default function makeRemoveQuestion({questionsDb}) {
         if (!id) {
             throw new Error('You must supply the question id.')
         }
-
+        if (!(id.match(/^[0-9a-fA-F]{24}$/))) {
+            throw new Error(`${id} is not a valid ObjectId`);
+        }
         return questionsDb.remove({id : id});
 
     }
